// src/app/general-processing/models/fee-detail.model.ts
export interface FeeDetail {
    accountEntityId: string;
    chargeType: string;
    dueAmount: string | number;
    feeAmount: string;
    feeName: string;
    paidAmount: string;
    payment: string;
    shoppingCartItemsId: string;
    statePlate: string;
    violationAccountChargesId: string;

}